[Start] follow_function

+++ app/controllers/book_comments_controller.rb
@@ -1,5 +1,6 @@
 class BookCommentsController < ApplicationController
-  
+before_action :authenticate_user!
+
   def create
     book = Book.find(params[:book_id])
     comment = BookComment.new(book_comment_params)
@@ -11,6 +12,11 @@ class BookCommentsController < ApplicationController
 
   def destroy
     comment = BookComment.find(params[:book_id])
+    # URLへの直リンクを禁止
+    if comment.user != current_user
+      # 1つ前のURLへ返す。
+      redirect_to request.referer
+    end
     comment.destroy
     redirect_back(fallback_location: root_path)
   end
